Hide symbols for GPU kernels (#310)
The -fvisibility=hidden flag does not affect kernels and thus all
__global__ GPU functions become externally-visible symbols by default.
Marking them as static will hide them.

This is not strictly necessary, since the rocrand kernels are all
appropriately confined to the rocrand details namespace. However, it
is still useful to hide their symbols, as it is difficult to review ABI
changes between different versions rocrand when the output of
dpkg-gensymbols is full of kernels that change from release to release.

[ROCm/rocRAND commit: 429bb626412e215ae380f9004f4f6a212e4566c8]