Remove build dependency on Git (#327)
The rocRAND project might use Git to download dependencies, but it is
not required if the dependencies are already installed. And, even in
the case that Git is needed to download dependencies, this find_package
call remains unnecessary. The tooling that does the downloads will look
for git itself.

This change was first made in 0d31194ab9811d9f19e78ec536529643c563401f,
but was reverted by f9db763c9c5c430d4a2c67639c25f6e44f02d134.

[ROCm/rocRAND commit: 896af30144d131383a1cd7442d83897d873dc004]