[windows] Find hipblaslt library with posix directory layout. (#2049)
Prior to this patch, hipblaslt would install its library under
lib/hipblaslt, which matches the Posix layout, but it was hard-coded to
only search adjacent to the hipblaslt.dll file (which on Windows, is in
the bin directory). For some reason, some existing Windows build logic
does expect to find it in the bin directory, and I am left to assume
that some out of tree patches exist that munge things into that form.
That behavior has been retained, but on Windows, we now first search the
sibling `lib` directory, assuming that the default layout of ROCM is the
same on all platforms (which is the case we are enforcing going
forward).

* The last time this code was touched, ellosel recommended that we
remove the library finding redundancy by introducing a helper.
* Refactored the three call sites to use a new common helper.
* Fixed up some raw string-based path handling that was requiring a '/'
in the path and defaulting silently to the PWD if not (!!!). This has
been causing many issues on Windows builds and people have been dumping
code objects into the bin directory to compensate.
* While writing the common utility, I moved the current standard
directory layout to the top of the search list (it was being searched
last, after old layouts).

Tested on Windows with TheRock.

[ROCm/hipBLASLt commit: 50d9abdddc0fda4b00313dfbc19e435d3b3a3500]