[next-cmake] Add new build system (#1629)
This is part one of a parallel bridge strategy that we are executing across multiple BLAS library components (based on work done in hipblaslt) to fix layering violations that cannot be safely addressed incrementally. It is an NFC until later in the process once it has been brought up in CI, remaining issues burned down, and a plan made to change the default.

---------

Co-authored-by: David Dixon <david.dixon@amd.com>